PLFM-5767: Added notification bean for SQS
